refactor: 136 - move products-sort into different folder
